% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_gombe_queries.R
\name{add_age}
\alias{add_age}
\title{Add age to a table}
\usage{
add_age(old_df, bio_tab = bio, idcol, datecol, newcolname = "age", deletebday = T)
}
\arguments{
\item{old_df}{Name of database to which you wish to add chimp ages}

\item{bio_tab}{Name of biography table from which to find birthdays. Defaults to bio.}

\item{idcol}{Name of column in old_df with chimp IDs. Use quotation marks around colname}

\item{datecol}{Name of column in old_df with date on which to calculate ages. Use quotes.}

\item{newcolname}{Name of new age column. Defaults to "age".}

\item{deletebday}{Whether to delete temporary birthday column. Defaults to TRUE.}
}
\value{
Returns the original table plus an age column.
}
\description{
Add age to a table
}
\examples{
bio = gombe_bio()
dais = data.frame(date = lubridate::dmy("1/1/2000"),
                   id1 = c("WL", "FN", "FF", "AO"),
                   id2 = rep("TB", times = 4),
                   dai = rnorm(n = 4, mean = 0, sd = 1))
dais = add_age(old_df = dais, idcol = "id1",
                datecol = "date", newcolname = "age1")
dais = add_age(old_df = dais, idcol = "id2",
                datecol = "date", newcolname = "age2")

#adds new column to table that has an age column for "id1", and another for "id2"

}
