% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dat.R
\name{gombe_a_rec}
\alias{gombe_a_rec}
\title{Import the A-Rec attendance table --KK ONLY--}
\usage{
gombe_a_rec(rename = TRUE, trim = TRUE)
}
\arguments{
\item{rename}{Whether to rename columns with friendlier names.  Default is TRUE.}

\item{trim}{Whether to remove community ID, observers, and old cycle info columns.
Default is TRUE.}
}
\value{
Returns the a_rec table.
}
\description{
Import the A-Rec attendance table --KK ONLY--
}
\examples{
a_rec = gombe_a_rec()
#imports the A-Record attendance table as "a_rec"

}
