% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dat.R
\name{gombe_agg_event}
\alias{gombe_agg_event}
\title{Import aggression_event table}
\usage{
gombe_agg_event(rename = TRUE, trim = FALSE)
}
\arguments{
\item{rename}{Whether to rename columns with friendlier names.  Default is TRUE.}

\item{trim}{Whether to remove full description, source, extracted by, and year columns.  Default is FALSE.}
}
\value{
Returns the aggression event table.
}
\description{
Import aggression_event table
}
\examples{
agg = gombe_agg_event()
#saves the aggression event table as "agg"

}
