% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_gombe_queries.R
\name{gombe_bio_id}
\alias{gombe_bio_id}
\title{Return subset of the biography table}
\usage{
gombe_bio_id(id)
}
\arguments{
\item{id}{ID desired (as individual or vector of IDs)}
}
\value{
Returns desired subset of biography table.
}
\description{
Return subset of the biography table
}
\examples{
gombe_bio_id(id = "FN")
#returns Fanni's life deets

}
