% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{gombe_bio_test}
\alias{gombe_bio_test}
\title{Import the biography table}
\usage{
gombe_bio_test(rename = TRUE)
}
\arguments{
\item{rename}{Whether to rename columns with friendlier names.  Default is TRUE, so set to false for
compatibility with old code.}
}
\value{
Returns the biography table.
}
\description{
Import the biography table
}
\examples{
bio = gombe_bio_test()
#saves the biography table as "bio"

}
