% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dat.R
\name{gombe_comm_memb}
\alias{gombe_comm_memb}
\title{Import the community_membership table}
\usage{
gombe_comm_memb(convert = TRUE, rename = TRUE)
}
\arguments{
\item{convert}{Whether you want to convert date columns to date format.  Default is TRUE.}

\item{rename}{Renames HK to KH and renames column headers to nicer names}
}
\value{
Returns the community_membership table.
}
\description{
Import the community_membership table
}
\examples{
comm = gombe_comm_memb()
#saves the community_membership table as "comm"

}
