% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_gombe_queries.R
\name{gombe_demog}
\alias{gombe_demog}
\title{Get demography at given date}
\usage{
gombe_demog(ddate = "31/12/2016", comms = "all")
}
\arguments{
\item{ddate}{date in form of DAY/MONTH/YEAR, <= "31/12/2016"}

\item{comms}{Communities desired.  Defaults to all, but can be in ("KK", "MT", "main", periph", "all")}
}
\value{
Returns demography counts on a given date
}
\description{
Get demography at given date
}
\examples{
demog = gombe_demog(ddate = "30/6/1986", comms = "KK")
  #returns demography counts for KK community on June 30, 1986

}
