% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dat.R
\name{gombe_fem_repro_hist}
\alias{gombe_fem_repro_hist}
\title{Import NEW female reproductive states table}
\usage{
gombe_fem_repro_hist(rename = TRUE)
}
\arguments{
\item{rename}{Whether to rename columns with friendlier names.  Default is TRUE, so set to false for
compatibility with old code.}
}
\value{
Returns the female reproductive states table.  NB: uses the newer
  reproductive STATES table, not the older reproductive history table
}
\description{
Import NEW female reproductive states table
}
\examples{
frh = gombe_fem_repro_hist()
#saves the female reproductive history table as "frh"

}
