% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dat.R
\name{gombe_fiss_comms}
\alias{gombe_fiss_comms}
\title{Import male post-fission communities table}
\usage{
gombe_fiss_comms()
}
\value{
Returns a table of communities males ended up in post 1973 fission.
}
\description{
Import male post-fission communities table
}
\examples{
fiss_com = gombe_fiss_comms()
#saves the aggression event table as "fiss_com"

}
