% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dat.R
\name{gombe_fol_arr}
\alias{gombe_fol_arr}
\title{Import follow_arrival table}
\usage{
gombe_fol_arr(rename = TRUE)
}
\arguments{
\item{rename}{Whether to rename columns with friendlier names.  Default is TRUE.}
}
\value{
Returns the follow arrival table.
}
\description{
Import follow_arrival table
}
\examples{
fol_arr = gombe_fol_arr()
#saves the aggression event table as "fol_arr"

}
