% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dat.R
\name{gombe_follow}
\alias{gombe_follow}
\title{Import follow table}
\usage{
gombe_follow()
}
\value{
Returns the follow table.
}
\description{
Import follow table
}
\examples{
fol = gombe_follow()
#saves the aggression event table as "fol"

}
