% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_secondary_dat.R
\name{gombe_food_lookup_tabs}
\alias{gombe_food_lookup_tabs}
\title{Import food lookup tables}
\usage{
gombe_food_lookup_tabs()
}
\value{
Returns the food_lookup, food_part_lookup, food_vars_lookup, and new_food_list tables.
  The tables are returned as a list, which can be separated by assigning list elements to
  new tables
}
\description{
Import food lookup tables
}
\examples{
food_lookup_tabs = gombe_food_lookup_tabs()
#imports the food lookup tables as list called "food_lookup_tabs"

#then to get food_part_lookup table as a data frame:
food_part_lookup = food_lookup_tabs$food_part_lookup

}
