% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dat.R
\name{gombe_pg_event}
\alias{gombe_pg_event}
\title{Import pantgrunt_event table}
\usage{
gombe_pg_event(rename = TRUE, trim = FALSE, cleaned = TRUE)
}
\arguments{
\item{rename}{Whether to rename columns with friendlier names.  Default is TRUE.}

\item{trim}{Whether to remove full notes, source, extracted by, and year columns.  Default is FALSE.}

\item{cleaned}{Whether to load pg table with multiple actors in individual rows rather than
one row separated by ommas}
}
\value{
Returns the pantgrunt event table.
}
\description{
Import pantgrunt_event table
}
\examples{
pg = gombe_pg_event()
#saves the aggression event table as "pg"

}
