% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_secondary_dat.R
\name{gombe_phenology}
\alias{gombe_phenology}
\title{Import phenology tables}
\usage{
gombe_phenology()
}
\value{
Returns the ripe_fruit, unripe_fruit, flowers, climate_data, and tree_locations tables.
  The tables are returned as a list, which can be separated by assigning list elements to
  new data frames
}
\description{
Import phenology tables
}
\examples{
phenology = gombe_phenology()
#imports the phenology as list called "phenology"

#then to get ripe_fruit table as a data frame:
ripe_fruit = phenology$ripe_fruit

}
