% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gombedatabase.R
\docType{package}
\name{gombedatabase}
\alias{gombedatabase}
\alias{gombedatabase-package}
\title{gombedatabase: The Gombe database in R}
\description{
The gombedatabase package is meant to do two things:
load database tables and run queries
}
\section{Thoughts on future development}{

\itemize{
  \item save data in /data/ folder and follow http://r-pkgs.had.co.nz/data.html
  \item save all data as .RData instead of csvs
  \item provide a mechanism for updating files (use instructions in above link)
  \item maybe add a function to check for new csv files, if it finds them, then update the .RData files?
  \item maybe add mindate, maxdate, and female_ID arguments to fem_repro_hist function
  \item Add an "add_community" function. But how? rely on date column in existing df? that's probs the right move
}
}

\section{Primary table import functions}{

\code{\link{gombe_bio}},
\code{\link{gombe_bio_test}}
\code{\link{gombe_comm_memb}},
\code{\link{gombe_fem_repro_hist}},
\code{\link{gombe_fem_repro_hist_NEW}},
\code{\link{gombe_agg_event}},
\code{\link{gombe_follow}},
\code{\link{gombe_fol_arr}},
\code{\link{gombe_fiss_comms}},
\code{\link{gombe_food_bout}},
\code{\link{gombe_a_rec}},
\code{\link{gombe_pg_event}}
}

\section{Common gombe DB queries}{

\code{\link{add_age}},
\code{\link{gombe_demog}},
\code{\link{gombe_bio_id}}
}

\section{Secondary table import functions}{

\code{\link{gombe_food_lookup_tabs}},
\code{\link{gombe_phenology}}
}

